+++ src/stylable.ts
@@ -3,7 +3,7 @@ import { FileProcessor, MinimalFS } from './cached-process-file';
 import { createInfrastructure } from './create-infra-structure';
 import { Diagnostics } from './diagnostics';
 import { safeParse } from './parser';
-import { StylableResolver } from './postcss-resolver';
+import { StylableResolver } from './stylable-resolver';
 import { process, StylableMeta } from './stylable-processor';
 import { Options, StylableResults, StylableTransformer, TransformHooks } from './stylable-transformer';
